Radios autofocussing again :D
